Refactor : change files structure
